new2

+++ lib/providers/background_provider.dart
@@ -3,7 +3,8 @@ import 'package:to_do_app/data/constants.dart' as constant;
 import 'package:to_do_app/utilities/app_shared_preferences.dart';
 
 class BackgroundProvider with ChangeNotifier, DiagnosticableTreeMixin {
-  late int _currentBackgroundIndex = AppSharedPreferences.getBackground();
+  late int _currentBackgroundIndex =
+      AppSharedPreferences.getBackground() as int;
   final bg = constant.wallpaper;
 
   int get currentBackgroundIndex => _currentBackgroundIndex;
